# prevent unnecessary `hash` collisions for `Tuple`