--- a/base/tuple.jl
+++ b/base/tuple.jl
@@ -578,11 +578,7 @@ end
 const tuplehash_seed = UInt === UInt64 ? 0x77cfa1eef01bca90 : 0xf01bca90
 function hash(tup::Tuple, h::UInt)
     @_terminates_locally_meta
-    f = let h = h
-        function closure_hash_tuple(e)
-            hash(e, h)
-        end
-    end
+    f = Fix2(hash, h)
     tuph = map(f, tup)
     out = hash(tuplehash_seed, h)
     for g = tuph