--- a/base/tuple.jl
+++ b/base/tuple.jl
@@ -577,6 +577,15 @@ end
 
 const tuplehash_seed = UInt === UInt64 ? 0x77cfa1eef01bca90 : 0xf01bca90
 const tuplehash_seed_final = UInt === UInt64 ? 0xf1bcd1f144b8c946 : 0x3c87fd81
+function tuplehash_fold_nonrecursive(tup::Tuple, h::UInt)
+    for e ∈ tup
+        h = hash(e, h)::UInt
+    end
+    h
+end
+tuplehash_fold(tup::Any32, h::UInt) = tuplehash_fold_nonrecursive(tup, h)  # don't give inference excessive work to do, `tup` is huge
+tuplehash_fold(tup::(Tuple{X, Vararg{X}} where {X}), h::UInt) = tuplehash_fold_nonrecursive(tup, h) # no need for recursion, `tup` is homogeneous
+tuplehash_fold(tup::All32, h::UInt) = tuplehash_fold_nonrecursive(tup, h)  # disambiguate
 tuplehash_fold(::Tuple{}, h::UInt) = h
 function tuplehash_fold(tup::Tuple{Any, Vararg}, h::UInt)
     f = first(tup)