Issue pointed out by foobar_lv2 on Discourse.