@nsajko 2025-02-09
Currently this PR is a slight performance regression. More importantly, it causes thousands of collisions here:

```julia
using Random

function experiment(generator)
    seed = rand(UInt32, 10)
    r = generator(seed)
    s = copy(r)
    collision_count = 0
    for j ∈ 1:1000000
        rand(r)
        collision_count += (hash(r) == hash(s))::Bool
        rand(s)
    end
    collision_count
end

function experiment_mt()
    experiment(MersenneTwister)
end

@show experiment_mt()
```

I'll rewrite the PR.