@nsajko 2025-02-09
## Benchmark

```julia
using InteractiveUtils: versioninfo
using BenchmarkTools
const data = (-10, -93, 116, 49, 38, 55, -120, -32, -14, -15, 83, 79, 67, 10, 113, 32)
const tuples_small = typeof(data)[(data .+ i) for i ∈ 1:10]
const tuples_large = typeof(data)[(data .+ i) for i ∈ 1:100]
const hs = [UInt(7)]
@btime hash(t, h) setup=(t = tuples_small; h = only(hs);)
@btime hash(t, h) setup=(t = tuples_large; h = only(hs);)
versioninfo()
```

It seems the PR now somehow results in a slight speedup.

### `master` branch

```none
  269.358 ns (0 allocations: 0 bytes)
  2.651 μs (0 allocations: 0 bytes)
Julia Version 1.13.0-DEV.24
Commit 79e98e3b89d (2025-02-09 00:36 UTC)
Build Info:
  Official https://julialang.org release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
  WORD_SIZE: 64
  LLVM: libLLVM-18.1.7 (ORCJIT, znver2)
  GC: Built with stock GC
Threads: 1 default, 1 interactive, 1 GC (on 8 virtual cores)
Environment:
  JULIA_NUM_PRECOMPILE_TASKS = 4
  JULIA_PKG_PRECOMPILE_AUTO = 0
```

### PR branch

```none
  258.058 ns (0 allocations: 0 bytes)
  2.549 μs (0 allocations: 0 bytes)
Julia Version 1.13.0-DEV.35
Commit b9b3a06bf48 (2025-02-09 02:35 UTC)
Build Info:
  Official https://julialang.org release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
  WORD_SIZE: 64
  LLVM: libLLVM-18.1.7 (ORCJIT, znver2)
  GC: Built with stock GC
Threads: 1 default, 1 interactive, 1 GC (on 8 virtual cores)
Environment:
  JULIA_NUM_PRECOMPILE_TASKS = 4
  JULIA_PKG_PRECOMPILE_AUTO = 0
```