@stevengj 2025-02-09
Would also be good to benchmark smaller tuples (e.g. 3-tuples) and heterogeneous or nested tuples.