@nsajko 2025-02-10
Actually, one thing I haven't tried is using `afoldl` directly, instead of `foldr`/`foldl`. Maybe that will work.