@nsajko 2025-02-11
## New benchmarks

Now we're never faster, but there's also no great regressions.

Benchmarking script:
```julia
function apply_offset(n::Int, offset::Int)
    (n + offset) % Int8
end
function apply_offset(t::Tuple, offset::Int)
    f = Base.Fix2(apply_offset, offset)
    map(f, t)
end

const tuple_16 = (-10, -93, 116, 49, 38, 55, -120, -32, -14, -15, 83, 79, 67, 10, 113, 32)
const tuple_03 = tuple_16[1:3]
const tuple_nested_heterogeneous = (
    tuple_16[1:2],
    (
        tuple_16[3],
        (
            tuple_16[4:6],
            tuple_16[7:8],
            tuple_16[9:10],
            tuple_16[11:13],
        ),
        (
            (
                tuple_16[14:15],
                (
                    tuple_16[16],
                ),
            ),
        ),
    ),
)

const hs = [UInt(7)]
const len = 20

const tuples_03 =
    [apply_offset(tuple_03, i) for i ∈ 1:len]
const tuples_16 =
    [apply_offset(tuple_16, i) for i ∈ 1:len]
const tuples_nested_heterogeneous =
    [apply_offset(tuple_nested_heterogeneous, i) for i ∈ 1:len]

using BenchmarkTools

for _ ∈ 1:3
    @btime hash(v, h) setup=(v = tuples_03; h = only(hs);)
    @btime hash(v, h) setup=(v = tuples_16; h = only(hs);)
    @btime hash(v, h) setup=(v = tuples_nested_heterogeneous; h = only(hs);)
    println()
end

using InteractiveUtils: versioninfo
versioninfo()
```

Results:
* `master` branch:
   ```none
     95.433 ns (0 allocations: 0 bytes)
     523.503 ns (0 allocations: 0 bytes)
     554.806 ns (0 allocations: 0 bytes)

     95.432 ns (0 allocations: 0 bytes)
     523.759 ns (0 allocations: 0 bytes)
     554.786 ns (0 allocations: 0 bytes)

     95.432 ns (0 allocations: 0 bytes)
     523.707 ns (0 allocations: 0 bytes)
     554.806 ns (0 allocations: 0 bytes)

   Julia Version 1.13.0-DEV.31
   Commit ac79b9f4137 (2025-02-11 02:54 UTC)
   Build Info:
     Official https://julialang.org release
   Platform Info:
     OS: Linux (x86_64-linux-gnu)
     CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
     WORD_SIZE: 64
     LLVM: libLLVM-18.1.7 (ORCJIT, znver2)
     GC: Built with stock GC
   Threads: 8 default, 1 interactive, 8 GC (on 8 virtual cores)
   Environment:
     JULIA_NUM_PRECOMPILE_TASKS = 4
     JULIA_PKG_PRECOMPILE_AUTO = 0
   ```
* PR branch:
   ```none
     97.996 ns (0 allocations: 0 bytes)
     560.568 ns (0 allocations: 0 bytes)
     591.894 ns (0 allocations: 0 bytes)

     97.996 ns (0 allocations: 0 bytes)
     560.514 ns (0 allocations: 0 bytes)
     599.726 ns (0 allocations: 0 bytes)

     97.996 ns (0 allocations: 0 bytes)
     560.565 ns (0 allocations: 0 bytes)
     599.719 ns (0 allocations: 0 bytes)

   Julia Version 1.13.0-DEV.47
   Commit 2648cb15c12 (2025-02-11 07:34 UTC)
   Build Info:
     Official https://julialang.org release
   Platform Info:
     OS: Linux (x86_64-linux-gnu)
     CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
     WORD_SIZE: 64
     LLVM: libLLVM-18.1.7 (ORCJIT, znver2)
     GC: Built with stock GC
   Threads: 8 default, 1 interactive, 8 GC (on 8 virtual cores)
   Environment:
     JULIA_NUM_PRECOMPILE_TASKS = 4
     JULIA_PKG_PRECOMPILE_AUTO = 0
   ```