@stevengj 2025-02-11
So about 8% slower.